edina/fieldtrip-cobweb-project#41: moved the code from widgets inside the ui, widgets will be for form elements
